+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/RecordMetaDataBuilder.java
@@ -1302,5 +1302,4 @@ public class RecordMetaDataBuilder implements RecordMetaDataProvider {
             super("Error converting from protobuf", cause);
         }
     }
-
 }